% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{setaALR}
\alias{setaALR}
\title{Additive Log-Ratio Transform}
\usage{
setaALR(counts, ref, pseudocount = 1)
}
\arguments{
\item{counts}{A matrix of counts.}

\item{ref}{The reference column, specified either by name or by index.}

\item{pseudocount}{Numeric. Added to avoid log(0). Default is 1.}
}
\value{
A list with:
\describe{
  \item{method}{A string noting \code{ALR_ref=<ref>}.}
  \item{counts}{A matrix of dimension \code{nrow(counts) x (ncol(counts) - 1)}.}
}
}
\description{
Applies an ALR transform to a matrix of counts using a specified reference column.
}
\details{
ALR transforms the data by taking \eqn{\log(x_i / x_{ref})} for each column
\eqn{i} except the reference column. This is another way to map data from
the simplex to a Euclidean space in compositional data analysis.
}
\examples{
#' mat <- matrix(c(1,2,4,8), nrow=2, byrow=TRUE)
colnames(mat) <- c("A", "B")
out <- setaALR(mat, ref="A", pseudocount=0)
out$counts

}
\references{
Aitchison, J. (1982). The Statistical Analysis of Compositional Data.
\emph{Journal of the Royal Statistical Society. Series B (Methodological)}, 44(2), 139-177.
}
