% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{setaCLR}
\alias{setaCLR}
\title{Centered Log-Ratio Transform}
\usage{
setaCLR(counts, pseudocount = 1)
}
\arguments{
\item{counts}{A matrix of counts (rows = features, columns = samples).}

\item{pseudocount}{Numeric.
Added to all entries to avoid taking log(0). Default is 1.}
}
\value{
A list with two elements:
\describe{
  \item{method}{A string indicating the transform ("CLR").}
  \item{counts}{A matrix of the same dimensions
                as the input after CLR transform.}
}
}
\description{
Applies a CLR transform to a matrix of counts.
This transform is commonly used in compositional data analysis (CoDA)
to project counts into a log-ratio space.
}
\details{
The CLR transform is defined as \eqn{\log(x / g(x))}
where \eqn{g(x)} is the geometric mean
of each row (sample) in the log scale.
A pseudocount helps avoid log(0) - default is 1, as scRNA data can be sparse.
}
\examples{
mat <- matrix(c(1,2,4,8), nrow=2, byrow=TRUE)
out <- setaCLR(mat, pseudocount=0)
out$counts

}
\references{
Aitchison, J. (1982). The Statistical Analysis of Compositional Data.
\emph{Journal of the Royal Statistical Society.
      Series B (Methodological)}, 44(2), 139-177.
}
