% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setaCounts}
\alias{setaCounts}
\title{Extract Taxonomic Counts from Various Single Cell Objects}
\usage{
setaCounts(obj, cell_type_col = "type", sample_col = "sample", bc_col = "bc")
}
\arguments{
\item{obj}{Either a \code{SingleCellExperiment}, a \code{Seurat} object, or
a \code{data.frame}. This function requires specific metadata or columns
to function correctly.}

\item{cell_type_col}{The name of the column representing cell types.
Default is "type".}

\item{sample_col}{The name of the column representing sample identifiers.
Default is "sample".}

\item{bc_col}{The name of the column representing barcodes (only needed for
\code{data.frame} input). Default is "bc".}
}
\value{
A matrix whose rows are samples and whose columns are cell types,
  with entries of the count of unique barcodes per type-sample combination.
}
\description{
Given a \code{SingleCellExperiment}, \code{Seurat}, or
long-form \code{data.frame}, creates a type-by-sample matrix of cell counts.
Users can specify the column names for cell types, samples, and barcodes.
}
\details{
\itemize{
  \item \strong{SingleCellExperiment} Reads \code{colData(obj)}
  for the specified \code{cell_type_col} and \code{sample_col}.
  \item \strong{Seurat} Uses \code{obj@meta.data} for the specified
  \code{cell_type_col} and \code{sample_col}.
  \item \strong{data.frame} Counts cells per specified \code{cell_type_col}
  and \code{sample_col}.
}
If the specified columns are missing, an error is thrown.
}
\examples{
# For a data.frame with custom column names:
set.seed(687)
df <- data.frame(
  barcode = paste0("cell", 1:10),
  cellType = sample(c("Tcell", "Bcell"), 10, TRUE),
  sampleID = sample(c("sample1","sample2"), 10, TRUE)
)
cmat <- setaCounts(df,
                   cell_type_col = "cellType",
                   sample_col = "sampleID",
                   bc_col = "barcode")
print(cmat)


}
