% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{setaPercent}
\alias{setaPercent}
\title{Percentage Transform}
\usage{
setaPercent(counts)
}
\arguments{
\item{counts}{A matrix of counts.}
}
\value{
A list with:
\describe{
  \item{method}{\code{"percent"}.}
  \item{counts}{A matrix of the same dimension, with each column summing to 100.}
}
}
\description{
Converts columns (samples) of a counts matrix to percentages of their respective column sums.
}
\details{
Simple re-scaling for compositional-like interpretation in percentages.
Useful for simplified Wilcoxon rank comparisons and such
}
\examples{
mat <- matrix(c(1,2,4,8), nrow=2, byrow=TRUE)
out <- setaPercent(mat)
out$counts

}
